erofs: fix potential overflow calculating xattr_isize
[ Upstream commit 1b3567a1969b26f709d82a874498c0754ea841c3 ]

Given on-disk i_xattr_icount is 16 bits and xattr_isize is calculated
from i_xattr_icount multiplying 4, xattr_isize has a theoretical maximum
of 256K (64K * 4).

Thus declare xattr_isize as unsigned int to avoid the potential overflow.

Fixes: bfb8674dc044 ("staging: erofs: add erofs in-memory stuffs")
Link: https://lore.kernel.org/r/20230414061810.6479-1-jefflexu@linux.alibaba.com

+++ fs/erofs/internal.h
@@ -339,7 +339,7 @@ struct erofs_vnode {
 	unsigned char data_mapping_mode;
 	/* inline size in bytes */
 	unsigned char inode_isize;
-	unsigned short xattr_isize;
+	unsigned int xattr_isize;
 
 	unsigned xattr_shared_count;
 	unsigned *xattr_shared_xattrs;
